requirements.txt
GDAL>=3.8.0
click>=8.1.0
PyYAML>=6.0